fix-configuration panel issues
