+++ META-INF/plugin.xml
@@ -99,5 +99,6 @@
 
     <extensions defaultExtensionNs="com.intellij">
         <!-- Add your extensions here -->
+        <applicationConfigurable instance="com.github.setial.intellijjavadocs.configuration.configurable.JavaDocConfigurable" groupId="tools" displayName="JavaDoc"/>
     </extensions>
 </idea-plugin>
