+++ src/com/github/setial/intellijjavadocs/configuration/JavaDocConfiguration.java
@@ -21,4 +21,5 @@ public interface JavaDocConfiguration {
     @Nullable
     JavaDocSettings getConfiguration();
 
+    void updateConfiguration(JavaDocSettings javaDocSettings);
 }
